refactor(app): Add folder for app files
