+++ guifaffinity-front/src/app/App.css
