+++ guifaffinity-front/src/components/header/header.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import App from "../../App";
+import App from "../../app/App";
 
 test("display logo", () => {
   render(<App />);
